Safari-611.1.10/LayoutTests/crypto/subtle/aes-import-jwk-key-export-jwk-key-expected.txt

Test importing a JWK AES-CBC key with legnth 128 and then export it in JWK format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS exportedJwkKey.kty is jwkKey.kty
PASS exportedJwkKey.k is jwkKey.k
PASS exportedJwkKey.alg is jwkKey.alg
PASS exportedJwkKey.key_ops is jwkKey.key_ops
PASS exportedJwkKey.ext is jwkKey.ext
PASS successfullyParsed is true

TEST COMPLETE
